requirements.txt
python-dotenv
sqlalchemy

streamlit